+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Added `--consider-allmeshes` CLI argument to force PGPatcher to consider patching meshes even if they do not show up in plugins
 - Fixed invalid shader permutation with anisotropic lighting and parallax together
 - Fixed "ignore" in the confirmation dialog scrolling the list back to the top
+- Removed "Fix Effect Lighting" patcher as this has been superceded by Community Shaders volumetric shadows
 
 ## [1.0.0] - 2026-02-23